Edited Zombie Powerups to show information to both teams
I was convinced that all powerups should show to both teams if they arent useless. Since zombies can play around human powers by seeing cooldown, humans should be allowed to do the same to zombies.
From now on, only hidden info will be useless items, items where there are so many that they will spam chat if not hidden, and items that aren't pistols for humans (so it doesnt mess with the hud trying to show 2 items on a human).